requirements.txt
Flask
gunicorn
requests
python-dotenv
flask-cors
Flask-CORS